Move ConfigLoader to file system subdirectory
